Play/Pause transition is triggering cascased state changes
